Don't force --tab-width and --use-tabs option when used specifies a config
SublimeJsPrettier uses the `--tab-width` and `--tab-size` options to `prettier` with the buffer's tab size and tab width from Sublime. For most users, this will be the auto-detected tab width and tab size, since auto-detection is the default. However sometimes Sublime guesses wrong, despite the file being formatted by `prettier` (this happens to me on this file: https://github.com/drew-gross/mpl/blob/master/api.ts).

When this happens the tab width specified in your prettier configuration is overridden by the incorrectly auto-detected tab width detected by Sublime, which causes the result of formatting using SublimeJsPrettier to not match the result of formatting using `prettier` on the command line.